+++ openghg/standardise/footprints/_paris.py
@@ -1,7 +1,6 @@
 import logging
 from pathlib import Path
 from collections import defaultdict
-from typing import DefaultDict, Dict, List, Optional, Union
 import warnings
 from xarray import Dataset
 
@@ -26,15 +25,15 @@ def parse_paris(
     model: str,
     inlet: str,
     species: str,
-    met_model: Optional[str] = None,
-    network: Optional[str] = None,
-    period: Optional[Union[str, tuple]] = None,
+    met_model: str | None = None,
+    network: str | None = None,
+    period: str | tuple | None = None,
     continuous: bool = True,
     high_spatial_resolution: bool = False,
     time_resolved: bool = False,
     high_time_resolution: bool = False,
     short_lifetime: bool = False,
-) -> Dict:
+) -> dict:
     """
     Read and parse input footprints data in "paris" format.
 
@@ -102,7 +101,7 @@ def parse_paris(
 
     # Need to read the metadata from the footprints and then store it
     # Do we need to chunk the footprints / will a Datasource store it correctly?
-    metadata: Dict[str, Union[str, float, List[float]]] = {}
+    metadata: dict[str, str | float | list[float]] = {}
 
     metadata["data_type"] = "footprints"
     metadata["site"] = site
@@ -181,7 +180,7 @@ def parse_paris(
     # TODO - remove this once assign_attributes has been refactored
     key = "_".join((site, domain, model, inlet))
 
-    footprint_data: DefaultDict[str, Dict[str, Union[Dict, Dataset]]] = defaultdict(dict)
+    footprint_data: defaultdict[str, dict[str, dict | Dataset]] = defaultdict(dict)
     footprint_data[key]["data"] = fp_data
     footprint_data[key]["metadata"] = metadata