+++ openghg/standardise/meta/_attributes.py
@@ -1,4 +1,5 @@
-from typing import Any, Dict, Hashable, List, Optional, Tuple, Union, cast
+from typing import Any, cast
+from collections.abc import Hashable
 import logging
 from xarray import Dataset
 from openghg.types import optionalPathType
@@ -18,14 +19,14 @@ logger.setLevel(logging.DEBUG)  # Have to set level for logger as well as handle
 
 
 def assign_attributes(
-    data: Dict,
-    site: Optional[str] = None,
-    network: Optional[str] = None,
-    sampling_period: Optional[Union[str, float, int]] = None,
+    data: dict,
+    site: str | None = None,
+    network: str | None = None,
+    sampling_period: str | float | int | None = None,
     update_mismatch: str = "never",
     site_filepath: optionalPathType = None,
     species_filepath: optionalPathType = None,
-) -> Dict:
+) -> dict:
     """Assign attributes to each site and species dataset. This ensures that the xarray Datasets produced
     are CF 1.7 compliant. Some of the attributes written to the Dataset are saved as metadata
     to the Datasource allowing more detailed searching of data.
@@ -51,6 +52,7 @@ def assign_attributes(
     Returns:
         dict: Dictionary of combined data with correct attributes assigned to Datasets
     """
+    from openghg.standardise.meta import sync_surface_metadata
 
     for _, gas_data in data.items():
         site_attributes = gas_data.get("attributes", {})
@@ -80,6 +82,19 @@ def assign_attributes(
             species_filepath=species_filepath,
         )
 
+        measurement_data = gas_data["data"]
+        metadata = gas_data["metadata"]
+
+        attrs = measurement_data.attrs
+
+        metadata_aligned, attrs_aligned = sync_surface_metadata(
+            metadata=metadata, attributes=attrs, update_mismatch=update_mismatch
+        )
+
+        gas_data["metadata"] = metadata_aligned
+        gas_data["attributes"] = attrs_aligned
+        measurement_data.attrs = gas_data["attributes"]
+
     return data
 
 
@@ -87,12 +102,12 @@ def get_attributes(
     ds: Dataset,
     species: str,
     site: str,
-    network: Optional[str] = None,
-    global_attributes: Optional[Dict[str, str]] = None,
-    units: Optional[str] = None,
-    scale: Optional[str] = None,
-    sampling_period: Optional[Union[str, float, int]] = None,
-    date_range: Optional[List[str]] = None,
+    network: str | None = None,
+    global_attributes: dict[str, str] | None = None,
+    units: str | None = None,
+    scale: str | None = None,
+    sampling_period: str | float | int | None = None,
+    date_range: list[str] | None = None,
     site_filepath: optionalPathType = None,
     species_filepath: optionalPathType = None,
 ) -> Dataset:
@@ -204,7 +219,7 @@ def get_attributes(
 
     ancillary_variables = []
 
-    variable_names = cast(Dict[str, Any], ds.variables)
+    variable_names = cast(dict[str, Any], ds.variables)
 
     # Write units as attributes to variables containing any of these
     match_words = ["variability", "repeatability", "stdev", "count"]
@@ -239,7 +254,7 @@ def get_attributes(
 
     # Add quality flag attributes
     # NOTE - I've removed the whitespace before status_flag and integration_flag here
-    variable_names = cast(Dict[str, Any], ds.variables)
+    variable_names = cast(dict[str, Any], ds.variables)
     quality_flags = [key for key in variable_names if "status_flag" in key]
 
     for key in quality_flags:
@@ -254,7 +269,7 @@ def get_attributes(
             "long_name": f"{long_name} status_flag",
         }
 
-    variable_names = cast(Dict[str, Any], ds.variables)
+    variable_names = cast(dict[str, Any], ds.variables)
     # Add integration flag attributes
     integration_flags = [key for key in variable_names if "integration_flag" in key]
 
@@ -271,7 +286,7 @@ def get_attributes(
 
     ds.time.encoding = {"units": f"seconds since {str(first_year)}-01-01 00:00:00"}
 
-    time_attributes: Dict[str, str] = {}
+    time_attributes: dict[str, str] = {}
     time_attributes["label"] = "left"
     time_attributes["standard_name"] = "time"
     time_attributes["comment"] = (
@@ -292,7 +307,7 @@ def get_attributes(
     return ds
 
 
-def define_species_label(species: str, species_filepath: optionalPathType = None) -> Tuple[str, str]:
+def define_species_label(species: str, species_filepath: optionalPathType = None) -> tuple[str, str]:
     """Define standardised label to use for observation datasets.
     This uses the data stored within openghg_defs/data/site_info JSON file
     by default with alternative names ('alt') defined within.
@@ -341,8 +356,8 @@ def define_species_label(species: str, species_filepath: optionalPathType = None
 
 
 def _site_info_attributes(
-    site: str, network: Optional[str] = None, site_filepath: optionalPathType = None
-) -> Dict:
+    site: str, network: str | None = None, site_filepath: optionalPathType = None
+) -> dict:
     """Reads site attributes from JSON
 
     This uses the data stored within openghg_defs/data/site_info JSON file by default.
@@ -395,13 +410,13 @@ def _site_info_attributes(
 
 
 def assign_flux_attributes(
-    data: Dict,
-    species: Optional[str] = None,
-    source: Optional[str] = None,
-    domain: Optional[str] = None,
+    data: dict,
+    species: str | None = None,
+    source: str | None = None,
+    domain: str | None = None,
     units: str = "mol/m2/s",
-    prior_info_dict: Optional[dict] = None,
-) -> Dict:
+    prior_info_dict: dict | None = None,
+) -> dict:
     """
     Assign attributes for the input flux dataset within dictionary based on
     metadata and passed arguments.
@@ -439,7 +454,7 @@ def assign_flux_attributes(
                 except KeyError:
                     raise ValueError(f"Attribute {attr} must be specified.")
 
-        input_attributes = cast(Dict[str, str], attribute_values)
+        input_attributes = cast(dict[str, str], attribute_values)
 
         flux_dict["data"] = get_flux_attributes(
             ds=flux_dict["data"],
@@ -458,8 +473,8 @@ def get_flux_attributes(
     source: str,
     domain: str,
     units: str = "mol/m2/s",
-    prior_info_dict: Optional[dict] = None,
-    global_attributes: Optional[Dict[Hashable, Any]] = None,
+    prior_info_dict: dict | None = None,
+    global_attributes: dict[Hashable, Any] | None = None,
 ) -> Dataset:
     """
     Assign additional attributes for the flux dataset.
@@ -508,7 +523,7 @@ def get_flux_attributes(
     ds["lon"].attrs = lon_attrs
 
     # Define default values for global attributes
-    global_attributes_default: Dict[Hashable, Any] = {
+    global_attributes_default: dict[Hashable, Any] = {
         "conditions_of_use": "Ensure that you contact the data owner at the outset of your project.",
         "Conventions": "CF-1.8",
     }
@@ -565,8 +580,8 @@ def get_flux_attributes(
 
 
 def dataset_formatter(
-    data: Dict,
-) -> Dict:
+    data: dict,
+) -> dict:
     """
     Formats species/variables from the dataset by removing the whitespaces
     with underscores and species to lower case
@@ -601,7 +616,7 @@ def data_variable_formatter(ds: Dataset, species: str, species_label: str) -> Da
     Returns:
         ds: xarray dataset
     """
-    variable_names = cast(Dict[str, Any], ds.variables)
+    variable_names = cast(dict[str, Any], ds.variables)
     to_underscores = {var: var.lower().replace(" ", "_") for var in variable_names}
     to_underscores.pop("time")  # Added to remove warning around resetting time index.
     ds = ds.rename(to_underscores)  # type: ignore
@@ -609,7 +624,7 @@ def data_variable_formatter(ds: Dataset, species: str, species_label: str) -> Da
     species_lower = species.lower()
     species_search = species_lower.replace(" ", "_")
 
-    variable_names = cast(Dict[str, Any], ds.variables)
+    variable_names = cast(dict[str, Any], ds.variables)
     matched_keys = [var for var in variable_names if species_search in var]
 
     # If we don't have any variables to rename, raise an error