+++ openghg/standardise/surface/_gcwerks.py
@@ -1,14 +1,11 @@
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple, Union
 from pandas import DataFrame
 
 from openghg.standardise.meta import dataset_formatter
 from openghg.types import optionalPathType
 
 
-def find_files(
-    data_path: Union[str, Path], skip_str: Union[str, List[str]] = "sf6"
-) -> List[Tuple[Path, Path]]:
+def find_files(data_path: str | Path, skip_str: str | list[str] = "sf6") -> list[tuple[Path, Path]]:
     """A helper file to find GCWERKS data and precisions file in a given folder.
     It searches for .C files of the format macehead.19.C, looks for a precisions file
     named macehead.19.precions.C and if it exists creates a tuple for these files.
@@ -54,17 +51,17 @@ def find_files(
 
 
 def parse_gcwerks(
-    filepath: Union[str, Path],
-    precision_filepath: Union[str, Path],
+    filepath: str | Path,
+    precision_filepath: str | Path,
     site: str,
     network: str,
-    inlet: Optional[str] = None,
-    instrument: Optional[str] = None,
-    sampling_period: Optional[str] = None,
-    measurement_type: Optional[str] = None,
+    inlet: str | None = None,
+    instrument: str | None = None,
+    sampling_period: str | None = None,
+    measurement_type: str | None = None,
     update_mismatch: str = "never",
     site_filepath: optionalPathType = None,
-) -> Dict:
+) -> dict:
     """Reads a GC data file by creating a GC object and associated datasources
 
     Args:
@@ -145,7 +142,7 @@ def parse_gcwerks(
     return gas_data
 
 
-def _check_site(filepath: Path, site_code: str, gc_params: Dict) -> str:
+def _check_site(filepath: Path, site_code: str, gc_params: dict) -> str:
     """Check if the site passed in matches that in the filename
 
     Args:
@@ -179,7 +176,7 @@ def _check_site(filepath: Path, site_code: str, gc_params: Dict) -> str:
     return site_code
 
 
-def _check_instrument(filepath: Path, gc_params: Dict, should_raise: bool = False) -> Union[str, None]:
+def _check_instrument(filepath: Path, gc_params: dict, should_raise: bool = False) -> str | None:
     """Ensure we have the correct instrument or translate an instrument
     suffix to an instrument name.
 
@@ -219,9 +216,9 @@ def _read_data(
     site: str,
     instrument: str,
     network: str,
-    gc_params: Dict,
-    sampling_period: Optional[str] = None,
-) -> Dict:
+    gc_params: dict,
+    sampling_period: str | None = None,
+) -> dict:
     """Read data from the data and precision files
 
     Args:
@@ -257,7 +254,7 @@ def _read_data(
         raise ValueError("Cannot process empty file.")
 
     # This metadata will be added to when species are split and attributes are written
-    metadata: Dict[str, str] = {
+    metadata: dict[str, str] = {
         "instrument": instrument,
         "site": site,
         "network": network,
@@ -282,7 +279,7 @@ def _read_data(
     units = {}
     scale = {}
 
-    flag_columns: List[Series] = []
+    flag_columns: list[Series] = []
     species = []
     columns_renamed = {}
     for column in data.columns:
@@ -352,7 +349,7 @@ def _read_data(
     return gas_data
 
 
-def _read_precision(filepath: Path) -> Tuple[DataFrame, List]:
+def _read_precision(filepath: Path) -> tuple[DataFrame, list]:
     """Read GC precision file
 
     Args:
@@ -388,12 +385,12 @@ def _split_species(
     data: DataFrame,
     site: str,
     instrument: str,
-    species: List,
-    metadata: Dict,
-    units: Dict,
-    scale: Dict,
-    gc_params: Dict,
-) -> Dict:
+    species: list,
+    metadata: dict,
+    units: dict,
+    scale: dict,
+    gc_params: dict,
+) -> dict:
     """Splits the species into separate dataframe into sections to be stored within individual Datasources
 
     Args:
@@ -535,12 +532,12 @@ def _split_species(
             combined_data[data_key]["data"] = spec_data
             combined_data[data_key]["attributes"] = attributes
 
-    to_return: Dict = combined_data.to_dict()
+    to_return: dict = combined_data.to_dict()
 
     return to_return
 
 
-def _get_sampling_period(instrument: str, gc_params: Dict) -> str:
+def _get_sampling_period(instrument: str, gc_params: dict) -> str:
     """Process the suffix from the filename to get the correct instrument name
     then retrieve the sampling period of that instrument from metadata.
 
@@ -560,7 +557,7 @@ def _get_sampling_period(instrument: str, gc_params: Dict) -> str:
     return sampling_period
 
 
-def _get_inlets(site_code: str, gc_params: Dict) -> Dict:
+def _get_inlets(site_code: str, gc_params: dict) -> dict:
     """Get the inlets we expect to be at this site and create a
     mapping dictionary so we get consistent labelling.
 
@@ -585,7 +582,7 @@ def _get_inlets(site_code: str, gc_params: Dict) -> Dict:
     return mapping_dict
 
 
-def _get_site_attributes(site: str, inlet: str, instrument: str, gc_params: Dict) -> Dict[str, str]:
+def _get_site_attributes(site: str, inlet: str, instrument: str, gc_params: dict) -> dict[str, str]:
     """Gets the site specific attributes for writing to Datsets
 
     Args:
@@ -601,7 +598,7 @@ def _get_site_attributes(site: str, inlet: str, instrument: str, gc_params: Dict
     site = site.upper()
     instrument = instrument.lower()
 
-    attributes: Dict[str, str] = gc_params["sites"][site]["global_attributes"]
+    attributes: dict[str, str] = gc_params["sites"][site]["global_attributes"]
 
     attributes["inlet_height_magl"] = format_inlet(inlet, key_name="inlet_height_magl")
     try: