+++ openghg/store/_obssurface.py
@@ -1,15 +1,16 @@
 from __future__ import annotations
 import logging
 from pathlib import Path
-from typing import Any, DefaultDict, Dict, Optional, Sequence, Tuple, Union, cast
+from typing import Any, cast
+from collections.abc import Sequence
 
 import numpy as np
 from pandas import Timedelta
 from xarray import Dataset
-from openghg.standardise.meta import sync_surface_metadata
 from openghg.store import DataSchema
 from openghg.store.base import BaseStore
 from openghg.types import multiPathType, pathType, resultsType, optionalPathType
+from collections import defaultdict
 
 logger = logging.getLogger("openghg.store")
 logger.setLevel(logging.DEBUG)  # Have to set level for logger as well as handler
@@ -26,11 +27,11 @@ class ObsSurface(BaseStore):
     def read_data(
         self,
         binary_data: bytes,
-        metadata: Dict,
-        file_metadata: Dict,
-        precision_data: Optional[bytes] = None,
+        metadata: dict,
+        file_metadata: dict,
+        precision_data: bytes | None = None,
         site_filepath: optionalPathType = None,
-    ) -> Dict:
+    ) -> dict:
         """Reads binary data passed in by serverless function.
         The data dictionary should contain sub-dictionaries that contain
         data and metadata keys.
@@ -108,14 +109,14 @@ class ObsSurface(BaseStore):
         source_format: str,
         site: str,
         network: str,
-        inlet: Optional[str] = None,
-        height: Optional[str] = None,
-        instrument: Optional[str] = None,
-        data_level: Optional[Union[str, int, float]] = None,
-        data_sublevel: Optional[Union[str, float]] = None,
-        dataset_source: Optional[str] = None,
-        sampling_period: Optional[Union[Timedelta, str]] = None,
-        calibration_scale: Optional[str] = None,
+        inlet: str | None = None,
+        height: str | None = None,
+        instrument: str | None = None,
+        data_level: str | int | float | None = None,
+        data_sublevel: str | float | None = None,
+        dataset_source: str | None = None,
+        sampling_period: Timedelta | str | None = None,
+        calibration_scale: str | None = None,
         measurement_type: str = "insitu",
         verify_site_code: bool = True,
         site_filepath: optionalPathType = None,
@@ -124,11 +125,11 @@ class ObsSurface(BaseStore):
         save_current: str = "auto",
         overwrite: bool = False,
         force: bool = False,
-        compressor: Optional[Any] = None,
-        filters: Optional[Any] = None,
-        chunks: Optional[Dict] = None,
-        optional_metadata: Optional[Dict] = None,
-    ) -> Dict:
+        compressor: Any | None = None,
+        filters: Any | None = None,
+        chunks: dict | None = None,
+        optional_metadata: dict | None = None,
+    ) -> dict:
         """Process files and store in the object store. This function
             utilises the process functions of the other classes in this submodule
             to handle each data type.
@@ -313,7 +314,7 @@ class ObsSurface(BaseStore):
                 filepath = fp
 
             # Cast so it's clear we no longer expect a tuple
-            filepath = cast(Union[str, Path], filepath)
+            filepath = cast(str | Path, filepath)
             filepath = Path(filepath)
 
             fn_input_parameters["filepath"] = filepath
@@ -364,8 +365,6 @@ class ObsSurface(BaseStore):
                 for key, value in data.items():
                     data[key]["data"] = value["data"].chunk(chunks)
 
-            self.align_metadata_attributes(data=data, update_mismatch=update_mismatch)
-
             # Check to ensure no required keys are being passed through optional_metadata dict
             # before adding details
             self.check_info_keys(optional_metadata)
@@ -403,7 +402,7 @@ class ObsSurface(BaseStore):
         measurement_type: str = "insitu",
         if_exists: str = "auto",
         overwrite: bool = False,
-    ) -> DefaultDict:
+    ) -> defaultdict:
         """Read AQMesh data for the Glasgow network
 
         NOTE - temporary function until we know what kind of AQMesh data
@@ -506,7 +505,7 @@ class ObsSurface(BaseStore):
 
         name = define_species_label(species)[0]
 
-        data_vars: Dict[str, Tuple[str, ...]] = {name: ("time",)}
+        data_vars: dict[str, tuple[str, ...]] = {name: ("time",)}
         dtypes = {name: np.floating, "time": np.datetime64}
 
         source_format = DataSchema(data_vars=data_vars, dtypes=dtypes)
@@ -534,14 +533,14 @@ class ObsSurface(BaseStore):
 
     def store_data(
         self,
-        data: Dict,
+        data: dict,
         if_exists: str = "auto",
         overwrite: bool = False,
         force: bool = False,
-        required_metakeys: Optional[Sequence] = None,
-        compressor: Optional[Any] = None,
-        filters: Optional[Any] = None,
-    ) -> Optional[Dict]:
+        required_metakeys: Sequence | None = None,
+        compressor: Any | None = None,
+        filters: Any | None = None,
+    ) -> dict | None:
         """This expects already standardised data such as ICOS / CEDA
 
         Args:
@@ -635,7 +634,7 @@ class ObsSurface(BaseStore):
 
         return datasource_uuids
 
-    def store_hashes(self, hashes: Dict) -> None:
+    def store_hashes(self, hashes: dict) -> None:
         """Store hashes of data retrieved from a remote data source such as
         ICOS or CEDA. This takes the full dictionary of hashes, removes the ones we've
         seen before and adds the new.
@@ -686,32 +685,3 @@ class ObsSurface(BaseStore):
 
     def set_hash(self, file_hash: str, filename: str) -> None:
         self._file_hashes[file_hash] = filename
-
-    def align_metadata_attributes(self, data: Dict, update_mismatch: str) -> None:
-        """
-        Function to sync metadata and attributes if mismatch is found
-
-        Args:
-            data: Dictionary of source_name data, metadata, attributes
-            update_mismatch: This determines how mismatches between the internal data
-                "attributes" and the supplied / derived "metadata" are handled.
-                This includes the options:
-                    - "never" - don't update mismatches and raise an AttrMismatchError
-                    - "from_source" / "attributes" - update mismatches based on input data (e.g. data attributes)
-                    - "from_definition" / "metadata" - update mismatches based on associated data (e.g. site_info.json)
-        Returns:
-            None
-        """
-        for _, gas_data in data.items():
-            measurement_data = gas_data["data"]
-            metadata = gas_data["metadata"]
-
-            attrs = measurement_data.attrs
-
-            metadata_aligned, attrs_aligned = sync_surface_metadata(
-                metadata=metadata, attributes=attrs, update_mismatch=update_mismatch
-            )
-
-            gas_data["metadata"] = metadata_aligned
-            gas_data["attributes"] = attrs_aligned
-            measurement_data.attrs = gas_data["attributes"]